# api/rest/docs/swagger.yaml
definitions:
...
  rest.BlockUpdateRequest:
    properties:
      badv:
        items:
          type: string
        type: array
      bcat:
        items:
          type: string
        type: array
      domain:
        type: string
      publisher:
        type: string
    type: object
  rest.BlockUpdateRespose:
    properties:
      status:
        description: 'in: body'
        type: string
    type: object
  rest.ConfiantUpdateRequest:
    properties:
      domain:
        type: string
      hash:
        type: string
      publisher:
        type: string
      rate:
        type: number
    type: object
  rest.ConfiantUpdateRespose:
    properties:
      status:
        description: 'in: body'
        type: string
    type: object
  rest.DemandFactorUpdateRequest:
    properties:
      demand_partner:
        type: string
      factor:
        type: number
    type: object
  rest.DemandFactorUpdateRespose:
    properties:
      status:
        description: 'in: body'
        type: string
    type: object
...
info:
  contact: {}
paths:
  /block:
    get:
      consumes:
      - application/json
      description: Get publisher block list (bcat and badv) setup
      produces:
      - text/html
      responses: {}
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
    post:
      consumes:
      - application/json
      description: Update bidder addomain and categories blocks.
      parameters:
      - description: Block update Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.BlockUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.BlockUpdateRespose'
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
...
  /confiant:
    get:
      consumes:
      - application/json
      description: Get confiant setup
      produces:
      - text/html
      responses: {}
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
    post:
      consumes:
      - application/json
      description: Update and enable Confiant setup (publisher is mandatory, domain
        is optional)
      parameters:
      - description: Confiant update Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.ConfiantUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.ConfiantUpdateRespose'
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
  /demand/factor:
    post:
      consumes:
      - application/json
      description: Update new bidder demand factor for demand partner.
      parameters:
      - description: DemandFactor update Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.DemandFactorUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.DemandFactorUpdateRespose'
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
...
swagger: "2.0"